+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add possibility to group tests based on delimiter in their names
+
 ## [0.9.0] - 2020-07-28
 
 ### Added