+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix issue #17 for tests with non-default `WORKING_DIRECTORY`
+
 ## [0.9.0] - 2020-07-28
 
 ### Added