+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.0] - 2020-08-28
+
 ### Added
 
-- Add possibility to group tests based on delimiter in their names
+- Add possibility to group tests based on delimiter in their names (thanks
+  @macdems!)
 
 ### Changed
 
@@ -89,7 +92,8 @@ Thanks to @andrewbridge for these contributions!
 
 - First release.
 
-[unreleased]: https://github.com/fredericbonnet/cmake-test-explorer/compare/v0.9.0...HEAD
+[unreleased]: https://github.com/fredericbonnet/cmake-test-explorer/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/fredericbonnet/cmake-test-explorer/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/fredericbonnet/cmake-test-explorer/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/fredericbonnet/cmake-test-explorer/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/fredericbonnet/cmake-test-explorer/compare/v0.6.0...v0.7.0